Rename and split-up utils.ts
