+++ packages/internal-utils/src/index.ts
@@ -3,7 +3,7 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 
+export * from "./common.js";
+export * from "./dom.js";
 export * from "./hooks.js";
-export * from "./icons.js";
-export * from "./symbol.js";
-export * from "./utils.js";
+export * from "./react.js";
